split command logic to separate file
